+++ APIBlox.NetCore.EventStore.EfCore/EfCoreSqlRepository.cs
@@ -2,12 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using APIBlox.NetCore.Contracts;
 using APIBlox.NetCore.Documents;
-using APIBlox.NetCore.Types;
 using Microsoft.EntityFrameworkCore;
 using Newtonsoft.Json;
 
@@ -22,34 +20,29 @@ namespace APIBlox.NetCore
         {
             _context = context;
         }
+
         public JsonSerializerSettings JsonSettings { get; }
 
         public async Task<int> AddAsync<TDocument>(TDocument[] documents, CancellationToken cancellationToken = default)
             where TDocument : EventStoreDocument
         {
             foreach (var document in documents)
-            {
-                var doc = JsonConvert.DeserializeObject<EventStoreDocumentEx>(JsonConvert.SerializeObject(document));
-
-                if (!(document.Data is null))
-                    doc.Data = JsonConvert.SerializeObject(document.Data);
-
-                _context.Documents.Add(doc);
-            }
+                _context.Documents.Add(document);
 
             var ret = await _context.SaveChangesAsync(cancellationToken);
 
             return ret;
         }
 
-        public async Task<IEnumerable<TResultDocument>> GetAsync<TResultDocument>(Expression<Func<IEventStoreDocument, bool>> predicate, CancellationToken cancellationToken = default)
-            where TResultDocument : IEventStoreDocument
+        public async Task<IEnumerable<TResultDocument>> GetAsync<TResultDocument>(Expression<Func<EventStoreDocument, bool>> predicate,
+            CancellationToken cancellationToken = default
+        )
+            where TResultDocument : EventStoreDocument
         {
             var ret = await _context.Documents.Where(predicate).ToListAsync(cancellationToken);
 
-            return (IEnumerable<TResultDocument>)ret;
+            return (IEnumerable<TResultDocument>) ret;
         }
-
 
         public async Task UpdateAsync<TDocument>(TDocument document, CancellationToken cancellationToken = default)
             where TDocument : EventStoreDocument
@@ -64,19 +57,16 @@ namespace APIBlox.NetCore
             await _context.SaveChangesAsync(cancellationToken);
         }
 
-        public async Task<int> DeleteAsync(Expression<Func<IEventStoreDocument, bool>> predicate, CancellationToken cancellationToken = default)
+        public async Task<int> DeleteAsync(Expression<Func<EventStoreDocument, bool>> predicate, CancellationToken cancellationToken = default)
         {
             var docs = await _context.Documents.Where(predicate).ToListAsync(cancellationToken);
 
             foreach (var doc in docs)
-            {
                 _context.Remove(doc);
-            }
 
             var ret = await _context.SaveChangesAsync(cancellationToken);
 
             return ret;
         }
     }
-
 }