+++ APIBlox.NetCore.EventStore.MongoDb/MongoDbRepository.cs
@@ -44,12 +44,11 @@ namespace APIBlox.NetCore
             return documents.Length;
         }
 
-        public async Task<IEnumerable<TResultDocument>> GetAsync<TResultDocument>(Expression<Func<IEventStoreDocument, bool>> predicate,
+        public async Task<IEnumerable<TResultDocument>> GetAsync<TResultDocument>(Expression<Func<EventStoreDocument, bool>> predicate,
             CancellationToken cancellationToken = default
-        )
-            where TResultDocument : IEventStoreDocument
+        )where TResultDocument : EventStoreDocument
         {
-            var ret = await _context.Collection<IEventStoreDocument>(_colName)
+            var ret = await _context.Collection<EventStoreDocument>(_colName)
                 .Find(predicate)
                 .ToListAsync(cancellationToken);
 
@@ -69,11 +68,11 @@ namespace APIBlox.NetCore
             );
         }
 
-        public async Task<int> DeleteAsync(Expression<Func<IEventStoreDocument, bool>> predicate,
+        public async Task<int> DeleteAsync(Expression<Func<EventStoreDocument, bool>> predicate,
             CancellationToken cancellationToken = default
         )
         {
-            var ret = await _context.Collection<IEventStoreDocument>(_colName)
+            var ret = await _context.Collection<EventStoreDocument>(_colName)
                 .DeleteManyAsync(predicate, null, cancellationToken);
 
             return (int)ret.DeletedCount;