+++ APIBlox.NetCore.EventStore/Contracts/IEventStoreRepository.cs
@@ -36,9 +36,9 @@ namespace APIBlox.NetCore.Contracts
         /// <param name="predicate">The predicate.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns>Task&lt;IEnumerable&lt;TResultDocument&gt;&gt;.</returns>
-        Task<IEnumerable<TResultDocument>> GetAsync<TResultDocument>(Expression<Func<IEventStoreDocument, bool>> predicate,
+        Task<IEnumerable<TResultDocument>> GetAsync<TResultDocument>(Expression<Func<EventStoreDocument, bool>> predicate,
             CancellationToken cancellationToken = default
-        )where TResultDocument : IEventStoreDocument;
+        )where TResultDocument : EventStoreDocument;
 
         /// <summary>
         ///     Updates the asynchronous.
@@ -55,6 +55,6 @@ namespace APIBlox.NetCore.Contracts
         /// <param name="predicate">The predicate.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns>Task&lt;System.Int32&gt;.</returns>
-        Task<int> DeleteAsync(Expression<Func<IEventStoreDocument, bool>> predicate, CancellationToken cancellationToken = default);
+        Task<int> DeleteAsync(Expression<Func<EventStoreDocument, bool>> predicate, CancellationToken cancellationToken = default);
     }
 }